+++ src/OpenVR.hpp
@@ -20,7 +20,7 @@ public:
         shutdown_vr();
     }
 
-    void init(IDirect3DDevice9* dev, const Config& cfg, IDirect3DVR9** vrdev, uint32_t companionWindowWidth, uint32_t companionWindowHeight);
+    void init(IDirect3DDevice9* dev, IDirect3DVR9** vrdev, uint32_t companionWindowWidth, uint32_t companionWindowHeight);
 
     void shutdown_vr() override
     {