+++ src/RenderTarget.hpp
@@ -11,6 +11,7 @@ enum RenderTarget : size_t {
 
 bool create_render_target(
     IDirect3DDevice9* dev,
+    D3DMULTISAMPLE_TYPE msaa,
     IDirect3DSurface9** msaa_surface,
     IDirect3DSurface9** depth_stencil_surface,
     IDirect3DTexture9** target_texture,
@@ -19,4 +20,4 @@ bool create_render_target(
     uint32_t w,
     uint32_t h);
 
-bool is_using_texture_to_render(RenderTarget t);
+bool is_using_texture_to_render(D3DMULTISAMPLE_TYPE msaa, RenderTarget t);